Allow installing salt with pytest-salt-factories by using extras.
This will allow projects, which are not Salt, but want to use the plugin
to depend on pytest-salt-factories and additionally pull in salt as a
dependency.
```
pip install pytest-salt-factories[salt]
```
Salt cannot be a hard dependency of the plugin because it's meant to
test salt in the first place, so, we can't bring a version of salt in
when we want to test another, likely unreleased, version of salt.

Fixes #22