Revert "Use salt's entry-points instead of relying on loader ``*_dirs`` configs"
This reverts commit 52050f6f9fce4cab24dcd13d2fe89be5aa13cefa.

Additionally, only inject ``engines_dirs`` and ``log_handlers_dirs`` when ``system_service=True`` or ``scripts_path`` is not ``None``.

These flags suggest that the salt being imported and used by salt-factories might not be the same as the one being tested.
So, in this case, make sure events and logging from started daemons still get forwarded to salt-factories.

Fixes #133